+++ okr-transparency-app/deploy/terraform/iam.tf
@@ -16,3 +16,9 @@ resource "google_project_iam_member" "api_secret_accessor" {
   role    = "roles/secretmanager.secretAccessor"
   member  = "serviceAccount:${google_service_account.api.email}"
 }
+
+resource "google_project_iam_member" "api_translation_user" {
+  project = var.project_id
+  role    = "roles/cloudtranslate.user"
+  member  = "serviceAccount:${google_service_account.api.email}"
+}